added first run
prevent the removal of the install dir  if upgrading

+++ entrypoint.sh
@@ -2,24 +2,30 @@
 
 printf "Checking for config_inc.php..."
 if [ -f "/config/config_inc.php" ]
-then 
-	chown www-data:www-data /config/config_inc.php
-	printf "Found\n"
-	printf "  Checking for instalation folder..."
-	if [ -d "/var/www/html/admin" ]
-	then 
-		printf "Found\n"
-		printf "    Atempting to remove..."
-		rm -rf /var/www/html/admin
-		if [ -d "/var/www/html/admin" ]
-		then
-			printf "Failed\n"
-		else
-			printf "Removed\n"
-		fi
-	else
-		printf "Missing - Good\n"
-	fi
+then
+	 if [ ! -f "/install-run" ]
+	 then
+  	 printf "creating tmpory install file for first fun\n"
+		 touch /install-run
+	 else
+		 chown www-data:www-data /config/config_inc.php
+		 printf "Found\n"
+		 printf "  Checking for instalation folder..."
+		 if [ -d "/var/www/html/admin" ]
+		 then
+		   printf "Found\n"
+		   printf "    Atempting to remove..."
+		   rm -rf /var/www/html/admin
+		   if [ -d "/var/www/html/admin" ]
+		   then
+  			 printf "Failed\n"
+  		 else
+  			 printf "Removed\n"
+  		 fi
+  	 else
+  		 printf "Missing - Good\n"
+  	 fi
+	 fi
 else
 	printf "Missing\n"
 	printf "Please use setup /admin/install.php \n"
